fixed logging and span issues with tests
