+++ raphtory-graphql/Cargo.toml
@@ -35,7 +35,7 @@ dynamic-graphql = { workspace = true }
 async-graphql-poem = { workspace = true }
 futures-util = { workspace = true }
 opentelemetry = { workspace = true }
-opentelemetry_sdk = { workspace = true }
+opentelemetry_sdk = { workspace = true, features = ["testing"] }
 opentelemetry-otlp = { workspace = true, features = ["http-json", "http-proto", "reqwest-rustls", "tls-aws-lc", "reqwest-blocking-client", "internal-logs"] }
 opentelemetry-stdout = { workspace = true }
 opentelemetry-appender-tracing = { workspace = true }
@@ -74,6 +74,8 @@ parquet = { workspace = true }
 opentelemetry_sdk = { version = "0.32.1", features = ["rt-tokio", "testing"] }
 
 [features]
+default = []
 python = ["dep:pyo3", "dep:pythonize", "raphtory/python"]
 search = ["raphtory/search"]
 panic-on-drop = ["raphtory/panic-on-drop"]
+integration-test = []
