+++ raphtory-graphql/src/config/app_config.rs
@@ -167,7 +167,7 @@ impl AppConfigBuilder {
                                         .map_err(|e| invalid_value([path, sub_path], e))?,
                                 );
                             }
-                            #[cfg(test)]
+                            #[cfg(any(feature = "integration-test", debug_assertions))]
                             TracingConfigFieldName::Exporters => {
                                 // Exporter handles are runtime-only and should not be loaded from config.
                             }
